GA supports both types of scores. Price and vrijstand calculation tweaked, both stored in building.
You can change the kind of score optimization the GA has to do with a
parameter. To make this more efficient price and vrijstand calculation
have been pulled apart. Besides that, for each house the price and
vrijstand can be stored which can be useful for adding more intelligence
in picking candidates in the GA.